sorting algo with vector is done !

+++ cpp09/ex02/JacobsthalNumberGenerator.hpp
@@ -6,7 +6,7 @@
 /*   By: hlahwaou <hlahwaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 08:56:40 by hlahwaou          #+#    #+#             */
-/*   Updated: 2024/01/25 10:00:39 by hlahwaou         ###   ########.fr       */
+/*   Updated: 2024/01/26 07:38:51 by hlahwaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ class JacobsthalNumberGenerator
 
     JacobsthalNumberGenerator();
     public:
-        JacobsthalNumberGenerator(size_t N);
+        JacobsthalNumberGenerator(int N);
         JacobsthalNumberGenerator(JacobsthalNumberGenerator const &);
         JacobsthalNumberGenerator& operator=(JacobsthalNumberGenerator const &);
         ~JacobsthalNumberGenerator();
